update benchmarks to new autodiff frontend
