+++ enzyme/benchmarks/ReverseMode/fft/src/lib.rs
@@ -1,6 +1,10 @@
 #![feature(slice_swap_unchecked)]
 #![feature(autodiff)]
-#![feature(slice_as_chunks)]
 
+#[cfg(not(any(feature = "safe", feature = "unsf")))]
+compile_error!("Enable at least one of: features `safe` or `unsf`");
+
+#[cfg(all(feature = "safe"))]
 pub mod safe;
+#[cfg(all(feature = "unsf"))]
 pub mod unsf;
