+++ enzyme/benchmarks/ReverseMode/lstm/src/lib.rs
@@ -1,10 +1,17 @@
 #![feature(autodiff)]
 
-pub (crate) mod unsf;
+#[cfg(not(any(feature = "safe", feature = "unsf")))]
+compile_error!("Enable at least one of: features `safe` or `unsf`");
+
+#[cfg(all(feature = "safe"))]
 pub (crate) mod safe;
+#[cfg(all(feature = "unsf"))]
+pub (crate) mod unsf;
+
 use std::slice;
 
 
+#[cfg(all(feature = "unsf"))]
 #[no_mangle]
 pub extern "C" fn rust_unsafe_lstm_objective(l: i32, c: i32, b: i32, main_params: *const f64, extra_params: *const f64, state: *mut f64, sequence: *const f64, loss: *mut f64) {
     let l = l as usize;
@@ -12,6 +19,7 @@ pub extern "C" fn rust_unsafe_lstm_objective(l: i32, c: i32, b: i32, main_params
     let b = b as usize;
     unsafe {unsf::lstm_unsafe_objective(l,c,b,main_params,extra_params,state,sequence, loss);}
 }
+#[cfg(all(feature = "safe"))]
 #[no_mangle]
 pub extern "C" fn rust_safe_lstm_objective(l: i32, c: i32, b: i32, main_params: *const f64, extra_params: *const f64, state: *mut f64, sequence: *const f64, loss: *mut f64) {
     let l = l as usize;
@@ -29,6 +37,7 @@ pub extern "C" fn rust_safe_lstm_objective(l: i32, c: i32, b: i32, main_params: 
     }
 }
 
+#[cfg(all(feature = "unsf"))]
 #[no_mangle]
 pub extern "C" fn rust_unsafe_dlstm_objective(l: i32, c: i32, b: i32, main_params: *const f64, d_main_params: *mut f64, extra_params: *const f64, d_extra_params: *mut f64, state: *mut f64, sequence: *const f64, res: *mut f64, d_res: *mut f64) {
     let l = l as usize;
@@ -36,6 +45,7 @@ pub extern "C" fn rust_unsafe_dlstm_objective(l: i32, c: i32, b: i32, main_param
     let b = b as usize;
     unsafe {unsf::d_lstm_unsafe_objective(l,c,b,main_params,d_main_params, extra_params,d_extra_params, state,sequence, res, d_res);}
 }
+#[cfg(all(feature = "safe"))]
 #[no_mangle]
 pub extern "C" fn rust_safe_dlstm_objective(l: i32, c: i32, b: i32, main_params: *const f64, d_main_params: *mut f64, extra_params: *const f64, d_extra_params: *mut f64, state: *mut f64, sequence: *const f64, res: *mut f64, d_res: *mut f64) {
     let l = l as usize;
